fix: fixing bugs and changing the project structure
