+++ src/entity/app/model/actions.js
@@ -1,0 +1,4 @@
+import { createAction } from 'redux-actions'
+
+export const INITIALIZE_APP = 'INITIALIZE_APP'
+export const initializeApp = createAction('INITIALIZE_APP')
