+++ src/entity/app/model/selector.js
@@ -1,0 +1,3 @@
+import { createSelector } from '@reduxjs/toolkit'
+
+export const selectApp = createSelector((state) => state.app.isInitialized)
