--- src/features/app/actions.js
@@ -1,5 +1,0 @@
-export const INITIALIZE_APP = 'INITIALIZE_APP'
-
-export const initializeApp = () => ({
-    type: INITIALIZE_APP
-})
