--- src/features/app/selector.js
@@ -1,3 +1,0 @@
-import { createSelector } from '@reduxjs/toolkit'
-
-export const selectApp = createSelector((state) => state.app.isInitialized)
