+++ src/store/index.js
@@ -1,1 +1,2 @@
 export { store } from './store'
+export { rootSaga } from './sagas'
