+++ src/store/reducers.js
@@ -1,7 +1,8 @@
 import { combineReducers } from 'redux'
 
-import { appReducer } from '../features/app/reducers'
-import { usersReducer } from '../features/users/reducers'
+import { appReducer } from '../entity/app'
+import { usersReducer } from '../entity/users'
+
 
 export const rootReducer = combineReducers({
     users: usersReducer,
