+++ src/store/store.js
@@ -2,7 +2,8 @@ import { configureStore } from '@reduxjs/toolkit'
 import createSagaMiddleware from 'redux-saga'
 
 import { rootReducer } from './reducers'
-import rootSaga from './sagas'
+import { rootSaga } from './sagas'
+
 
 const sagaMiddleware = createSagaMiddleware()
